Changed the buffer point to a float, and wrote more of the FFT function

+++ Application/SpectralViewer/Source/MainComponent.h
@@ -120,10 +120,10 @@ private:
                                         // Within the project.
     playState state;                    // State of viewer: Playing, Starting, Stopping, Stopped, or Paused.
     
-    double  projectTime;                // Current time at the source file.
-    double  fileLength;                 // Length of the project in seconds.
-    double  sampleFreq;                 // Sample rate of the selected file.
-    double* buffer;         // Will be used to point to an array storing samples.
+    double projectTime;                 // Current time at the source file.
+    double fileLength;                  // Length of the project in seconds.
+    double sampleFreq;                  // Sample rate of the selected file.
+    float* buffer;                      // Will be used to point to an array storing samples.
     
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (MainComponent)
